moved plugins in plugins folders, added only exports in index file
