+++ source-code/plugin-registry/src/registry.ts
@@ -1,0 +1,28 @@
+import type { InlangPluginManifest } from "./schema.js"
+
+export const plugins: InlangPluginManifest[] = [
+	{
+		id: "inlang-plugin-json",
+		icon: "https://github.com/samuelstroschein/inlang-plugin-json/assets/icon.png",
+		repository: "https://github.com/samuelstroschein/inlang-plugin-json",
+		keywords: ["json", "i18next", "svelte-intl-precompile"],
+		contributorHandle: "samuelstroschein",
+		contributorImage: "https://avatars.githubusercontent.com/u/35429197?v=4",
+	},
+	{
+		id: "plugin-i18next",
+		icon: "https://github.com/samuelstroschein/inlang-plugin-json/assets/icon.png",
+		repository: "https://github.com/samuelstroschein/inlang-plugin-json",
+		keywords: ["json", "i18next", "svelte-intl-precompile"],
+		contributorHandle: "nilsjacobsen",
+		contributorImage: "https://avatars.githubusercontent.com/u/35429197?v=4",
+	},
+	{
+		id: "inlang-lint",
+		icon: "https://github.com/samuelstroschein/inlang-plugin-json/assets/icon.png",
+		repository: "https://github.com/samuelstroschein/inlang-plugin-json",
+		keywords: ["json", "i18next", "svelte-intl-precompile"],
+		contributorHandle: "felixhaeberle",
+		contributorImage: "https://avatars.githubusercontent.com/u/35429197?v=4",
+	},
+]
